Trigger an error when hotfix is not installed or password is wrong
Hotfix XS65ESP1013 is to fix Cinder volume when iSCSI is used. It
is necessary to check whether this hotfix is installed and exit
deployment if not. And docs are also updated to make sure user will
do so in advance.

Besides, if password of XenServer is wrong, an alert with error
message will also be needed.

Change-Id: I2b7d5b1b127e8a0e232537a4baa5f401f9be0854
Closes-Bug: #1508474
Closes-Bug: #1508996