Fix novnc proxy configuration
The guest consoles are not visible through the horizon UI because
novnc proxy is set to compute node by default thus a port forwarding
to dom0 is required. Besides, Mirantis uses internal host name -
"public.fuel.local" - as the novnc's base url which cannot be
resolved by users therefore need to be replaced with the real public
IP address.

Change-Id: I819d13dca1a8263247fe443606095ba33b142756
Closes-Bug: #1513551