plugin-xenserver: Apply patch to speed up writing config drive
This is to apply a patch to address the following problem:
bug: #1567857
When writing config drive, dd by default uses 512B as the block
size. It costs more than 3mins to write the 64M config drive.
It's longer than expected: Some health checking(fuel's heat
checking) on XenServer may failed due to timeout. So we need
improve the performance via specifying a bigger block size
with dd.

Change-Id: I97e121800567efbce66ce1de8f8cbf9c788d12e5