Replace static-routes with udev rules
As stacic route table won't be able to work before the HIMN is brought up,
so a host reboot will make static route rules lost. But udev rules can do
the same thing when Compute node is connected to HIMN and the interface
xenapi will be up then.

As found there might be a time gap between udev script is running and
interface xenapi is ready, a 5-second sleep is added just in case.

Change-Id: Ifa3260107265f400bde2a33640d4da2fc9f32ddf