Replace static-routes with udev rules
When host is restarted, static-routes won't be able add route rules
because HIMN isn't brought up. Thus we need to use udev to add
those route rules back. As tested there is time gap between udev is
triggered and HIMN is ready, a 5-sec sleep is added just in case.

Change-Id: Ifa3260107265f400bde2a33640d4da2fc9f32ddf