Enable ceilometer
In order to enable ceilometer, need add xenapi specific setup in
ceilometer's configure file and remove it from the impatible list.
Note ceilometer is an option service, the users can choose if
enable it or not. This patch also contain a patch to include a fix
on polling cpu_util.

Change-Id: I0b60494aa1d4511b18d6582f15d454af43b8500c