Make branding based on the branch, to remove the PLUGIN_REVISION variable
By checking which branch we are on, we can automatically determine the
revision based on the length of history between the root of the branch
and this change.

Normally (and in development environments) this isn't guaranteed as you can
create arbitrary branches (e.g. branch a and branch b could have the same
merge base) but if run against code checked out from github will give a
consistent version number

Change-Id: I6eff6befce1408dac144dd7dfb3f3055820960a8