Check hotfixes for given platform
Hotfixes are platform specific, thus needed to be specified in branding.inc.
And hotfixes were hardcoded and now should be configurable. Previously
plugin binary might be compatible across some platform, but it will make
the hotfixes hard to configure. Here will try to identify platform version
first and fail the deployment if it is different with the specified value.

Change-Id: Ifb753c3cab3618471a8b642d42ecf9a815e3da28