Check for HIMN using dhcp discovery
A future plan for this plugin is to not require the HIMN XC
plugin but to support adding the network just in XenCenter
(with hidden objects displayed). Therefore the fuel plugin
should try to auto detect the HIMN as the MAC would not be
written to XenStore.

Cherry-pick of 01402e857c99a3a8aa76535cde81c479e330a0a6

Change-Id: I83aa26b16f9839979b63616bdf358219ff23739f