Fix novnc cache
In MOS 9 the memcache isn't enabled for consoleauth by default and for a
multi-controller environment the authentication will be failed because token
cannot be retrieved from memcache.

This fix also moves the deprecated settings from DEFAULT section to vnc or cache.
And the nova-compute.conf on controller actually never been used thus is removed.

Change-Id: Ic9438172b6c1ed5074aaa0b6d0e036287efe1982