Fix execute and file patching
The old execute() will silently ignore error regardless if it is in
allowed_return_codes. And the old file patching will get return code 1 when
running for the second time. The new implementation will firstly check if the
base file can be reversed(by dry-run). If it is so, then the patch will not be
applied.

Change-Id: I055428ec4b142ac9caa9b21a1609e2718e06aaf2
(cherry picked from commit 10874d8672c9d34b971ae174c10c58e6fe33de14)